Move StorageSail to sparql package since it is not related to enrichment
