+++ src/main/scala/se/lu/nateko/cp/meta/services/sparql/magic/EnrichingSail.scala
@@ -7,12 +7,11 @@ import org.eclipse.rdf4j.common.order.StatementOrder
 import org.eclipse.rdf4j.model.{IRI, Resource, Statement, Value, ValueFactory}
 import org.eclipse.rdf4j.query.Dataset
 import org.eclipse.rdf4j.query.algebra.evaluation.EvaluationStrategy
-import org.eclipse.rdf4j.query.algebra.evaluation.EvaluationStrategyFactory
 import org.eclipse.rdf4j.query.algebra.evaluation.TripleSource
 import org.eclipse.rdf4j.query.algebra.evaluation.impl.{DefaultEvaluationStrategy, DefaultEvaluationStrategyFactory, EvaluationStatistics}
-import org.eclipse.rdf4j.query.algebra.evaluation.federation.{FederatedServiceResolver, FederatedServiceResolverClient}
+import org.eclipse.rdf4j.query.algebra.evaluation.federation.FederatedServiceResolver
 import org.eclipse.rdf4j.sail.helpers.{NotifyingSailConnectionWrapper, NotifyingSailWrapper}
-import org.eclipse.rdf4j.sail.{NotifyingSail, NotifyingSailConnection}
+import org.eclipse.rdf4j.sail.NotifyingSailConnection
 import org.slf4j.LoggerFactory
 import se.lu.nateko.cp.meta.services.citation.{CitationClient, CitationProvider}
 
@@ -20,13 +19,11 @@ import scala.concurrent.{ExecutionContext, Future}
 import scala.reflect.Selectable.reflectiveSelectable
 import scala.util.{Failure, Success}
 
+import se.lu.nateko.cp.meta.services.sparql.StoreSail
 import se.lu.nateko.cp.meta.core.data.EnvriConfigs
 
 import StatementsEnricher.StatIter
-
-
 type MainSail = FederatedServiceResolverClient & NotifyingSail:
-	def setEvaluationStrategyFactory(factory: EvaluationStrategyFactory): Unit
 
 
 class EnrichingSail(
